[PR rtl-optimization/122321][RISC-V] Bounds check another access to ira_reg_equiv array
So another case where we're indexing into the ira_reg_equiv array without
checking bounds.  I sincerely hope we're not playing wack-a-mole here, but two
failures in a couple months for the same core problem is worrisome.

Bootstrapped and regression tested on x86_64 and riscv64 as well as run through
all the embedded targets in my tester without regressions.

	PR rtl-optimization/122321
gcc/
	* lra-constraints.cc (update_equiv): Make sure REGNO is in
	ira_reg_equiv before trying to update ira_reg_equiv.

gcc/testsuite/
	* gcc.target/riscv/rvv/autovec/pr122321.c: New test.