LoongArch: Eliminate unnecessary dependencies introduced by xvpermi.q
1. When the selector is 0x0, 0x1, 0x10, or 0x11, the result of
xvpermi.q does not depend on the output operand, thus eliminating
the dependency chain of the output operand as input, which can
reduce the number of instructions.

2. When the selector is 0x22, 0x23, 0x32, or 0x33, the result of
xvpermi.q does not depend on the second input operand, thus
eliminating the dependency chain of the second input operand,
which can also reduce the number of instructions.

gcc/ChangeLog:

	* config/loongarch/lasx.md (lasx_xvpermi_q_<LASX:mode>):
	Add new splitter for optimization.

gcc/testsuite/ChangeLog:

	* gcc.target/loongarch/vec_pack_unpack_256.c: Adjust to changed
	lasx_xvpermi_q_<LASX:mode> template.
	* gcc.target/loongarch/vector/lasx/lasx-builtin.c: Ditto.
	* gcc.target/loongarch/lasx-xvpermi_q-opt.c: New test.