Use gather_imm_use_stmts instead of FOR_EACH_IMM_USE_STMT in forwprop
The following fixes forwprop using FOR_EACH_IMM_USE_STMT to iterate
over stmts and then eventually removing the active stmt, releasing
its defs.  This can cause debug stmt insertion with a RHS referencing
the SSA name we iterate over, adding to its immediate use list
but also adjusting all other debug stmts refering to the released
SSA name, updating those.  And those can refer to the original
iterated over variable.

In the end the destructive behavior of update_stmt is a problem
here, which unlinks all uses of a stmt and then links in the
newly computed ones instead of leaving those in place that are.

The solution is to not rely on FOR_EACH_IMM_USE_STMT to iterate
over stmt uses without duplicates.

	* tree-ssa-forwprop.cc (forward_propagate_addr_expr):
	Use gather_imm_use_stmts instead of FOR_EACH_IMM_USE_STMT.