c++: modules, clones, and overload resolution
21_strings/basic_string_view/cons/wchar_t/3.cc was failing with import std
because as in the reduced testcase, the module includes an instantiation of
the template constructor for <const char *, const char *>, and importing the
module was wrongly adding that instantiation to CLASSTYPE_MEMBER_VEC, so it
became part of the overload set independent of its template.
tsubst_function_decl doesn't add to CLASSTYPE_MEMBER_VEC, and importing
should work the same.

gcc/cp/ChangeLog:

	* module.cc (trees_in::decl_value): Don't add an instantiation to
	CLASSTYPE_MEMBER_VEC.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/clone-5_a.C: New test.
	* g++.dg/modules/clone-5_b.C: New test.