RISC-V: testsuite: Fix pr119115.c.
pr119115.c is broken (since it was added?) but because we only
attempt to compile/run it on a 256b target I never noticed it.

This patch adds -std=gnu99 to make it compile.

Regtested on rv64gcv_zvl512b.  Going to push after the CI run.

Regards
 Robin

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/base/pr119115.c: Add -std=gnu99.