doc, cris: Clean up CRIS option documentation [PR122243]
This is another patch in the series to make documentation of
target-specific options in invoke.texi match what is in the
corresponding .opt files.

The cris.opt file is a bit strange, with many cases where negative
forms are listed explicitly as separate options from the positive
forms, with both having "RejectNegative" and one (typically the form
that is the default) being marked as "Undocumented".  I've left that
alone since fixing it to the more normal style of having a single
option setting a boolean flag would require code changes, and I'm not
set up to build or test this target.

Beyond that, the "Undocumented" status of options in the .opt file did
not in several cases match what was actually documented in the manual.
I've generally assumed that the manual is correct, and e.g. the
-m32-bit, -m16-bit, and -m8-bit options, all previously marked
"Undocumented" but listed in invoke.texi, are preferred to the
equivalent options without the hyphen.  I've removed the references to
the obsolete -melf and -maout options and added documentation in the
manual for some options that were previously documented only in the .opt
file.

gcc/ChangeLog
	PR other/122243
	* config/cris/cris.opt (m32-bit, m16-bit, m8-bit): Remove
	Undocumented property.
	(m32bit, m8bit): Add Undocumented property.
	* doc/invoke.texi (Option Summary) <CRIS Options>: Remove
	obsolete -melf and -maout options from table, plus redundant
	-mno-mul-bug-workaround.
	(CRIS Options): Add @opindex for -mno- forms that didn't already
	have one.  Remove obsolete -melf documentation.  Document
	-mbest-lib-options, -moverride-best-lib-options,
	-mtrap-using-break8, -mtrap-unaligned-atomic, and
	-munaligned-atomic-may-use-library.