arc: Fix positive form of -mno-brcc and -mno-dpfp-lrsr options [PR122243]
These two options are defined in the negative form and don't have the
RejectNegative property so they end up having positive forms beginning
with "-mno-no-", which is confusing.  I've inverted the sense of the
option instead (so that the positive forms are -mbrcc and
-mdpfp-lrsr).

I'm not set up to build or test gcc for ARC but this is a straightforward
change similar to fixes I've made for other target-inspecific options.
Either this fix or adding RejectNegative would break any makefiles
that use the (undocumented) -mno-no forms, though.

gcc/ChangeLog
	PR other/122243
	* config/arc/arc.opt (-mno-brcc, -mno-dpfp-lrsr): Redefine in
	the positive sense.