tree-optimization/122573 - enhance SLP of invariant loads
Currently SLP of invariant loads is only supported for the case of
a single load that is splat, as side-effect of supporting this case
even for non-invariant loads.  The following extends this to any
set of invariant loads.  The way we have load permutations for
these makes it a bit awkward, thus adjustments in that area.

	PR tree-optimization/122573
	* tree-vect-slp.cc (vect_build_slp_tree_1): Support
	groups of invariant loads.
	(vect_build_slp_tree_2): Likewise.
	(vect_transform_slp_perm_load_1): Likewise.
	* tree-vect-stmts.cc (vectorizable_load): Handle non-splat
	SLP for invaraint loads.

	* gcc.dg/vect/slp-58.c: New testcase.