[x86] avoid using masked vector epilogues when no scalar epilog is needed
The following arranges for avoiding masked vector epilogues when we'll
eventually arrive at a vector epilogue with VF == 1 which implies no
scalar epilog will be necessary.

This avoids regressing performance in OpenColorIO when the
avx512_masked_epilogues tuning is enabled.  A testcase for one
example case is shown in PR122573.

	PR tree-optimization/122573
	* config/i386/i386.cc (ix86_vector_costs::finish_cost): Avoid
	using masked epilogues when an SSE epilogue would have a VF of one.

	* gcc.dg/vect/costmodel/x86_64/costmodel-pr122573.c: New testcase.