doc, arm: Clean up ARM option documentation [PR122243]
This patch undocuments ARM target-specific options that have never
been implemented, are already marked as "Undocumented" in arm.opt
file, and/or are presently documented as obsolete or only useful for
back end debugging.  I've also cleaned up the option summary to list
only one of the positive or negative forms of each option, and to
consistently index both forms.

gcc/ChangeLog
	PR other/122243
	* config/arm/arm.opt (mapcs-reentrant): Mark as "Undocumented",
	updatehelp string for internal documentation.
	(mapcs-stack-check): Likewise update help string.
	(mprint-tune-info, mneon-for-64bits): Mark as "Undocumented".
	* doc/invoke.texi (Option Summary) <ARM Options>: Remove duplicate
	entries for negative forms and entries for options that are
	explicitly "Undocumented".  Add missing entry for
	-mpic-data-is-text-relative.  Fix some formatting issues.
	(ARM Options): Remove documentation for -mapcs-stack-check,
	-mapcs-reentrant, -mflip-thumb, -mneon-for-64-bits,
	-mprint-tune-info, and -mverbose-cost-dump.  Add index entries
	for -mno- option forms.  Minor editing for clarity.